+++ src/module/action.sh
@@ -1,5 +1,5 @@
 #!/system/bin/sh
-WORK_DIR="/data/adb/sshcustom"
+WORK_DIR="/data/adb/sshcustom-vpnchain"
 RUN_DIR="$WORK_DIR/run"
 CONTROL="$WORK_DIR/sshcustom.sh"
 LOG="$RUN_DIR/action.log"